mysite/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
Django==4.1.5
gevent==22.10.2
greenlet==2.0.1
idna==3.4
pycparser==2.21
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
websocket-client==1.4.2
zope.event==4.6
zope.interface==5.5.2
gunicorn==20.1.0
